Dialog fields are reset after submitted
The input fields for the new book are now empty when you click on the button again after adding a book.
The function to add the new book won't be started if any of the fields are empty. Alert will popup.
Changed "read" of the Book prototype to a boolean, to facilitate future changes.